requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.52
langchain-text-splitters==0.0.2
llama_cpp_python==0.2.75
pydantic==2.7.1
requests==2.31.0
sentence-transformers==2.7.0
streamlit==1.34.0
streamlit-chat==0.1.1
pypdf
fastembed
docx2txt
python-pptx
python-dotenv